requirements.txt
streamlit>=1.36.0
PyPDF2>=3.0.1
google-generativeai>=0.7.0
pytest>=8.2.0
